refactor: use kewyword args for gate parameters
The parsing technique which is currently under development transforms
the parameters given to the gate in the OpenQASM syntax into a simple
tuple of numbers; for these parameters to be turned into named
parameters, we would need to know the signature of each
gate function in the library, which would probably result in a gigantic
if-then-else block. It is best to just make each gate function accept
the additional parameters as normal parameters at the end of the
signature, and just make sure that we write them in order every time.