fix: replace Gell-Mann basis with PTM basis
Use the Pauli-transfer matrix basis to compute the coefficients of
matrices and (super-)operators, instead of the Gell-Mann basis.
It is very likely that they contain the same elements, but the
construction of the noise map in the sparse Pauli-Lindblad model
requires a particular order of the basis elements, so it is safer to
write it this way.