refactor: add generic `gate` method
Add a `gate` method that, mimicking ITensors' own `op` and `state`
functions, can be overloaded through a parametric type, which we call
`GateName`. By taking the `SiteType` as an argument, we can write
methods with the same name for both `Qubit` and `vQubit` site types and
let Julia dispatch at compile time.
The syntax of the gate functions was also simplified a bit, now using
`Index` objects directly instead of a `Vector{<:Index}` and some integer
indices to denote the qbit sites.

Note that in the transition to a Julia package the OpenQASM 2.0
gate set has beed dropped, since we have (for now) no method of
distinguishing them from the 3.0 ones in a way that is meaningful within
Julia's multiple dispatch.