test: check if same coefficient order as Qiskit
Add a small test script that loads an OpenQASM2.0 circuit generated by
a Qiskit script with the same name, and checks if the result is the
same, down to the coefficient order. The Python script must be run by
the user in some way before the Julia test script is executed, so that
the relevant files are generated.