test: try parsing more complicated circuit
Add a new test script which tries to emulate a somewhat complicated
Qiskit circuit, with two-qbit gates and more importantly U gates, which
differ between the OpenQASM 2.0 and 3.0 specifications, so that the most
important gate definitions are checked.
This also renders the previous qbit-ordering test useless, since if the
basis order is not correct this script will fail anyway.