test: write standardized and modular test scripts
Rewrite tests in the proper Julia way, i.e. in a `@testset` macro within
the `test/runtests.jl` file, so that testing can be automated and easily
performed by launching `] test` from the Julia REPL.
Also split the previous test in three more modular scripts:

1. a test of the qbit ordering
2. a test of the definition of the generic SU(2) gate
3. a test of the register-to-siteinds mapping

so that it's easier to see which part of the package doesn't behave
correctly in case of any issue.