feat: new gate library according to Qiskit defs
Add yet another gate library file, this time following Qiskit's
definitions, which include a SU(2) generic gate (the "U gate") which
despite what they claim in their documentation is not the same as the
one in the OpenQASM 3.0 specification.
Use this library as default, and leave the other two commented out in the
module entry point.