feat: compute noise and inverse-noise layers
Add a function that computes the MPO of a noise layer, starting from
either a set of parameters or a SPLNoiseModel object.
This is basically the same as the already existing function that
computed the inverse of the noise layer, so this previous code got a bit
optimized in the process.