fix: reverse PauliString only in string repr.
Instead of reversing the order of the terms when creating the ITensors
operator from the string, reverse the order only when converting to/from
the explicit String representation, i.e. the sequence of I, X, Y and Z
characters, which is how the package interfaces with Qiskit. This way
the internal order of the qbits in this package is not touched.