feat: `gate` automatically determines site types
Move the "classical arguments" of the gates in a keyword argument so
that `gate` can imitate ITensor's `op` find-and-dispatch mechanism.
This is better since the site type can be automatically inferred from
the given Index objects and does not have to be passed explicitly
in the function call anymore.
This is necessary to allow defining new gates on-the-fly from an
OpenQASM instruction, using the `definition` function, without having to
worry about the site types.