feat: use kwargs in `gate` only when necessary
Slightly change how `parsegate` is defined, so that if the list of
classical arguments is empty then `gate` is called without the `cargs`
keyword argument. If the gate (as read from the OpenQASM code) does not
have classical arguments, then, our ITensor version does not need to
accept keyword arguments. This way the `gate` methods can be written
more cleanly.