fix: `apply(a, b)` with no common indices
The `apply` function from ITensor multiplies tensors only if they have
common indices. A new gate might be defined such that tensors acting on
different qbits are multiplied together, and this would result in
ITensor objects with different indices that cannot be multiplied
together with `apply`. To solve this, define a `compose` function that
calls `apply` or the simple product `*` depending on whether the tensors
have or don't have common indices.