fix: parse gates based on existing ITensor sites
Add a new method for the `gates` function that parses and creates new
gates basing them on an existing list of ITensor sites.
This is useful when two or more circuits (that is, read from different
OpenQASM files) should be composed; the function cannot create another
set of sites or the gates of the different circuits wouldn't be
cross-compatible.