fix: don't forget to add last gate layer
At the end of the "layer creation" loop in the `gatelayers` function,
add the current layer to the list of layers.
Previously, the last gate was filled, but the line of code which
loaded the layer into the "complete layers" list was never reached, so
it was never returned by the function.