feat: embed `qbitmap` information in siteinds
Include the "name" of the qbits as in the OpenQASM code, i.e. "q[0]",
"q[1]" etc. directly in the tags of the site Index objects. This removes
the need for an external function that maps the original circuit qbit
names to the elements of the ITensor site indices.
As a consequence, the cumbersome `qbitmap` mechanism can be removed,
greatly simplifying the gate creation and calling methods, that do not
need the `qbitmap` arguments anymore.