perf: cache results of Pauli string contraction
Use the @memoize macro to cache the results of the contraction of
single Pauli string factors with an MPS. This saves a lot of time and
memory when the same MPS has to be contracted with a lot of different
Pauli strings, e.g. when computing overlaps in the `samplepaulistrings`
function.